multi_tool_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-generator==1.10
async-lru==2.1.0
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.8
Automat==25.4.16
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==1.0.0a4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
crashtest==0.3.1
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
Faker==38.2.0
fastapi==0.127.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
google-adk==1.25.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.137.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.1
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.36.1
google-cloud-storage==3.4.1
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.63.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
hiredis==2.4.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
incremental==24.7.2
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
lark==1.3.1
librt==0.7.5
lightgbm==4.6.0
llvmlite==0.46.0b1
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
more-itertools==10.8.0
MouseInfo==0.1.3
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.2
notebook==7.5.3
notebook_shim==0.2.4
numba==0.63.0b1
numpy==2.3.5
openai==2.20.0
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.6
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==9.0.21
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pptree==3.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.3.0
py-cpuinfo==9.0.0
pyarrow==23.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pylev==1.4.0
Pyment==0.3.3
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==9.0.1
pytest-benchmark==5.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.0.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-slugify==8.0.4
python-ulid==1.1.0
pytokens==0.3.0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
redis==4.6.0
redis-om==0.3.5
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
rubicon-objc==0.5.3
scikit-learn==1.7.2
scipy==1.16.3
Scrapy==2.13.3
seaborn==0.13.2
selenium==4.40.0
Send2Trash==2.1.0
sentry-sdk==2.46.0
service-identity==24.2.0
setuptools==80.9.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlalchemy-spanner==1.17.2
sqlmodel==0.0.31
sqlparse==0.5.4
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.4
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
timescaledb==0.0.4
tinycss2==1.4.0
tldextract==5.3.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
Twisted==25.5.0
typer==0.20.0
types-certifi==2021.10.8.3
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-requests==2.32.4.20250913
types-setuptools==80.9.0.20251223
types-six==1.17.0.20251009
types-tqdm==4.67.0.20250809
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uv==0.9.13
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.35.3
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
wsproto==1.3.2
xgboost==3.1.2
yarl==1.22.0
zipp==3.23.0
zope.interface==8.1.1
requests
python-dotenv
